backend/pyproject.toml
[project]
name = "codewright-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "bleach>=6.2.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "markdown-it-py>=4.0.0",
    "passlib>=1.7.4",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pygments>=2.19.2",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "rq>=2.4.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "weasyprint>=66.0",
]
